Fix inner indentation with index
idx -> (inc idx)

According to the README, the index specified for inner indentation is
the index of the argument.

For example, specifying an index of 0 means that the inner indentation
is targeting the first argument. However, the first argument does not
have an index of 0 within the form. It has an index of 1, because the
first argument should be the second expression in the current list.

(dec depth) -> depth

Considering that '(first (iterate z/up zloc))' is 'zloc' itself, it is
not necessary to decrement 'depth'.

Fixes #73.